requirements.txt
Flask
scikit-learn
pandas
numpy
tensorflow
keras
joblib
Pillow
